+++ CHANGELOG.md
@@ -13,6 +13,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Handle no payment result case. (#46)
 
+### Changed
+
+- Improve admin labels for deleted transactions. (#46)
+
 ## [1.8.2]
 
 ### Fixed